Cryptography Codes
My Solutions provides two security services;
-	A Secure Channel: My solution provides a secure channel using asymmetric encryption for credential/key exchange. This channel is based on RSA encryption which is a popular asymmetric encryption used for implementing public-private key encryption. This channel ensures that the secret exchange between the sender and the receiver is secure and can only be accessed with the receiver’s private key. 
-	An Insecure Channel: My Solution also provides an insecure channel that can handle large data, which is used to send the encrypted message/data from the sender to the receiver. The message is encrypted using symmetric encryption with AES. The secret used to encrypt the message is then sent to the receiver via the secure channel.
-	Data Integrity
-	Data confidentiality 
-	
What cryptographic algorithm has been used to encrypt the data? What cipher modes, if applicable, have you used? Why is this algorithm appropriate to the task?

I used the AES encryption algorithm to encrypt the message/data, the encrypted data is then sent to the receiver via an insecure channel. Inasmuch as it’s an insecure channel, we can confidently send data through it because the data has been encrypted using the AES encryption algorithm using the ECB cipher mode and the secret used to encrypt this data has been sent to the receiver via a secure channel implemented with RSA which is an asymmetric encryption algorithm meaning that only the receiver’s private key can decrypt the secret key and thereafter decrypt the message.
What other cryptographic algorithms have been used in your application? Again, why are these algorithms appropriately used in your application?

I used the RSA algorithm to implement a secure channel for the exchange of small data such as credentials, secrets, etc. In this use case, this channel was used to send the message encryption secret to the receiver so as to enable the receiver decrypt the message successfully. I used the PKCSI-OAEP cipher mode for the encryption mainly because this cipher mode adds padding to the RSA algorithm which basically makes it more secure. 
What are the key lengths for cryptographic keys? Why are these key lengths appropriate?

 The Secret for the AES encryption is 32-character string. Which I think delivers a relatively high encryption while also deliver an efficient encryption.
The key size for the RSA encryption is 2048bits which is the recommended key size for RSA encryptions. This key size delivers a high-level encryption while being relatively fast in terms of encryption speed.

Conclusion 

The AES key encryption method was used to encrypt the message in a secure channel while the RSA based secure channel was used to send the secret key to the receiver. Moreover, when Alice wants to send a message to Bob, Alice makes a request to get Bob’s public key, generates a random secret then use this secret to encrypt the message using the AES. In retrospect, I used the receiver’s public key to encrypt the secret using the RSA the sent the encrypted message with the encrypted secret (JAMESDYER.JPG) to the receiver. This message can’t be decrypted without the secret key.